feat: enhance restaurant and category services with error handling and new fields

+++ src/services/categorias_platos.services.js
@@ -37,7 +37,14 @@ export async function crearCategoriaPlato(categoria) {
 
 export async function eliminarCategoriaPlato(id) {
     try {
+        const categoria = await obtenerBD().collection(COLECCION_CATEGORIAS_PLATOS).findOne({ id: id });
+
+        if (!categoria) {
+            throw new Error("Categoría de plato no encontrada");
+        }
+
         return await obtenerBD().collection(COLECCION_CATEGORIAS_PLATOS).deleteOne({ id: id });
+
     } catch (error) {
         throw new Error("Error al eliminar la categoría de plato: " + error.message);
     }
@@ -45,10 +52,23 @@ export async function eliminarCategoriaPlato(id) {
 
 export async function actualizarCategoriaPlato(id_plato, id_categoriaActualizada) {
     try {
+
+        const plato = await obtenerBD().collection(COLECCION_PLATOS).findOne({ id: id_plato });
+        const categoria = await obtenerBD().collection(COLECCION_CATEGORIAS_PLATOS).findOne({ id: id_categoriaActualizada });
+
+        if (!plato) {
+            throw new Error("Plato no encontrado");
+        }
+
+        if (!categoria) {
+            throw new Error("Categoría de plato no encontrada");
+        }
+
         return await obtenerBD().collection(COLECCION_PLATOS).updateOne(
             { id: id_plato },
             { $set: { categoriaId: id_categoriaActualizada } }
         );
+        
     } catch (error) {
         throw new Error("Error al actualizar la categoría del plato: " + error.message);
     }
